ajout des tests unitaires pour DeletedTest, ajout des tests instrumentés pour TestDao, avec la javaDoc

+++ app/src/main/java/com/example/ex_motricite/DeletedTest.java
@@ -3,11 +3,11 @@ package com.example.ex_motricite;
 import androidx.annotation.NonNull;
 
 public class DeletedTest {
-    private int id;
+    private long id;
     private final String path;
     private String suppressionDate;
 
-    public DeletedTest(int id, String path, String suppressionDate) {
+    public DeletedTest(long id, String path, String suppressionDate) {
         this.id = id;
         this.path = path;
         this.suppressionDate = suppressionDate;
@@ -18,11 +18,11 @@ public class DeletedTest {
         this.suppressionDate = suppressionDate;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
