+++ app/src/test/java/com/example/ex_motricite/DeletedTestTest.java
@@ -9,7 +9,7 @@ class DeletedTestTest {
     //constructor
     void testDeletedTestWhenAllParamsGiven() {
         //GIVEN
-        int id = 1;
+        long id = 1;
         String path = "/directory/file";
         String suppressionDate = "01/01/2000";
         //WHEN
@@ -21,12 +21,12 @@ class DeletedTestTest {
     @Test
     void id_are_correctly_set() {
         // GIVEN
-        int id = 1;
+        long id = 1;
         String path = "/directory/file";
         String suppressionDate = "01/01/2000";
         com.example.ex_motricite.DeletedTest test = new DeletedTest(id, path, suppressionDate);
         // WHEN
-        int retrievedId = test.getId();
+        long retrievedId = test.getId();
         // THEN
         assertEquals(retrievedId, id);
     }
@@ -34,11 +34,11 @@ class DeletedTestTest {
     @Test
     void test_set_id_correctly_set() {
         // GIVEN
-        int id = 1;
+        long id = 1;
         String path = "/directory/file";
         String suppressionDate = "01/01/2000";
         com.example.ex_motricite.DeletedTest test = new DeletedTest(id, path, suppressionDate);
-        int newId = 2;
+        long newId = 2;
         // WHEN
         test.setId(newId);
         // THEN
@@ -48,7 +48,7 @@ class DeletedTestTest {
     @Test
     void test_path_is_correctly_retrieved() {
         // GIVEN
-        int id = 1;
+        long id = 1;
         String path = "/directory/file";
         String suppressionDate = "01/01/2000";
         com.example.ex_motricite.DeletedTest test = new DeletedTest(id, path, suppressionDate);
@@ -61,7 +61,7 @@ class DeletedTestTest {
     @Test
     void suppression_date_are_correctly_retrieved() {
         // GIVEN
-        int id = 1;
+        long id = 1;
         String path = "/directory/file";
         String suppressionDate = "01/01/2000";
         com.example.ex_motricite.DeletedTest test = new DeletedTest(id, path, suppressionDate);
@@ -74,12 +74,11 @@ class DeletedTestTest {
     @Test
     void test_set_suppression_date_correctly_apply() {
         // GIVEN
-        int id = 1;
+        long id = 1;
         String path = "/directory/file";
         String suppressionDate = "01/01/2000";
         com.example.ex_motricite.DeletedTest test = new DeletedTest(id, path, suppressionDate);
         String newSuppressionDate = "02/01/2000";
-
         // WHEN
         test.setSuppressionDate(newSuppressionDate);
         // THEN
